refactor: altering the path of the types
