+++ painel_administrativo-front/src/app/components/table_divergencia.tsx
@@ -2,7 +2,7 @@
 
 import { useRouter } from "next/navigation"
 
-import type { divergenciasType } from "../dashboard/types/types"
+import type { divergenciasType } from "../../types/types"
 
 import {
   formatToDDMMYYYYHHMM,
